+++ plugins/module_utils/network/waveserverai/facts/facts.py
@@ -21,9 +21,26 @@ from ansible_collections.ciena.waveserverai.plugins.module_utils.network.waveser
 )
 
 
+from ansible_collections.ciena.waveserverai.plugins.module_utils.network.waveserverai.facts.system.system import (
+    SystemFacts,
+)
+
+from ansible_collections.ciena.waveserverai.plugins.module_utils.network.waveserverai.facts.ptps.ptps import (
+    PtpsFacts,
+)
+
+from ansible_collections.ciena.waveserverai.plugins.module_utils.network.waveserverai.facts.ports.ports import (
+    PortsFacts,
+)
+
 FACT_LEGACY_SUBSETS = dict(default=Default, config=Config)
 
-FACT_RESOURCE_SUBSETS = dict(xcvrs=XcvrsFacts)
+FACT_RESOURCE_SUBSETS = dict(
+    ports=PortsFacts,
+    ptps=PtpsFacts,
+    xcvrs=XcvrsFacts,
+    system=SystemFacts,
+)
 
 
 class Facts(FactsBase):
@@ -35,9 +52,7 @@ class Facts(FactsBase):
     def __init__(self, module):
         super(Facts, self).__init__(module)
 
-    def get_facts(
-        self, legacy_facts_type=None, resource_facts_type=None, data=None
-    ):
+    def get_facts(self, legacy_facts_type=None, resource_facts_type=None, data=None):
         """Collect the facts for waveserverai
 
         :param legacy_facts_type: List of legacy facts types
@@ -48,13 +63,9 @@ class Facts(FactsBase):
         """
         # netres_choices = FactsArgs.argument_spec['gather_network_resources'].get('choices', [])
         if self.VALID_RESOURCE_SUBSETS:
-            self.get_network_resources_facts(
-                FACT_RESOURCE_SUBSETS, resource_facts_type, data
-            )
+            self.get_network_resources_facts(FACT_RESOURCE_SUBSETS, resource_facts_type, data)
 
         if self.VALID_LEGACY_GATHER_SUBSETS:
-            self.get_network_legacy_facts(
-                FACT_LEGACY_SUBSETS, legacy_facts_type
-            )
+            self.get_network_legacy_facts(FACT_LEGACY_SUBSETS, legacy_facts_type)
 
         return self.ansible_facts, self._warnings