+++ plugins/module_utils/network/waveserverai/facts/ptps/ptps.py
@@ -4,7 +4,7 @@
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 """
-The waveserverai xcvrs fact class
+The waveserverai ptps fact class
 It is in this file the configuration is collected from the device
 for a given resource, parsed, and the facts tree is populated
 based on the configuration.
@@ -28,8 +28,8 @@ from ansible_collections.ansible.netcommon.plugins.module_utils.network.common i
 from ansible_collections.ansible.netcommon.plugins.module_utils.network.netconf.netconf import (
     get,
 )
-from ansible_collections.ciena.xcvrs.plugins.module_utils.network.waveserverai.argspec.xcvrs.xcvrs import (
-    XcvrsArgs,
+from ansible_collections.ciena.ptps.plugins.module_utils.network.waveserverai.argspec.ptps.ptps import (
+    PtpsArgs,
 )
 
 try:
@@ -42,12 +42,12 @@ except ImportError:
     HAS_LXML = False
 
 
-class XcvrsFacts(object):
-    """The waveserverai xcvrs fact class"""
+class PtpsFacts(object):
+    """The waveserverai ptps fact class"""
 
     def __init__(self, module, subspec="config", options="options"):
         self._module = module
-        self.argument_spec = XcvrsArgs.argument_spec
+        self.argument_spec = PtpsArgs.argument_spec
         spec = deepcopy(self.argument_spec)
         if subspec:
             if options:
@@ -60,7 +60,7 @@ class XcvrsFacts(object):
         self.generated_spec = utils.generate_dict(facts_argument_spec)
 
     def populate_facts(self, connection, ansible_facts, data=None):
-        """Populate the facts for xcvrs
+        """Populate the facts for ptps
         :param connection: the device connection
         :param ansible_facts: Facts dictionary
         :param data: previously collected conf
@@ -72,15 +72,15 @@ class XcvrsFacts(object):
 
         if not data:
             config_filter = """
-                <waveserver-xcvrs xmlns="urn:ciena:params:xml:ns:yang:ciena-ws:ciena-waveserver-xcvr">
-                </waveserver-xcvrs>
+                <waveserver-ptps xmlns="urn:ciena:params:xml:ns:yang:ciena-ws:ciena-waveserver-ptp">
+                </waveserver-ptps>
                 """
             data = get(self._module, filter=("subtree", config_filter))
 
         stripped = remove_namespaces(xml_to_string(data))
         data = fromstring(to_bytes(stripped, errors="surrogate_then_replace"))
 
-        resources = data.xpath("/rpc-reply/data/waveserver-xcvrs/xcvrs")
+        resources = data.xpath("/rpc-reply/data/waveserver-ptps/ptps")
         objs = []
         for resource in resources:
             if resource:
@@ -90,10 +90,10 @@ class XcvrsFacts(object):
 
         facts = {}
         if objs:
-            facts["xcvrs"] = []
+            facts["ptps"] = []
             params = utils.validate_config(self.argument_spec, {"config": objs})
             for cfg in params["config"]:
-                facts["xcvrs"].append(utils.remove_empties(cfg))
+                facts["ptps"].append(utils.remove_empties(cfg))
 
         ansible_facts["ansible_network_resources"].update(facts)
         return ansible_facts