+++ plugins/modules/waveserverai_facts.py
@@ -94,13 +94,8 @@ def main():
 
     :returns: ansible_facts
     """
-    module = AnsibleModule(
-        argument_spec=FactsArgs.argument_spec, supports_check_mode=True
-    )
-    warnings = [
-        "default value for `gather_subset` "
-        "will be changed to `min` from `!config` v2.11 onwards"
-    ]
+    module = AnsibleModule(argument_spec=FactsArgs.argument_spec, supports_check_mode=True)
+    warnings = ["default value for `gather_subset` " "will be changed to `min` from `!config` v2.11 onwards"]
 
     result = Facts(module).get_facts()