+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Ciena WaveserverAi Collection Release Notes
 
-## v1.0.3
+## v1.1.0
+### Added
+- Added support for Waveserver Ai 2.5.0
+- System Module
+- Xcvrs Module
+- Ports Module
+- Ptps Module
 
+## v1.0.3
 ### Fixed
 - Moved config_is_diff method to utils
 
 ## v1.0.0
-
 ### Added
 - Initial